Started converting to strtoks. sscanf can't handle all the different GPS message styles.

+++ include/conversions.h
@@ -14,7 +14,7 @@
 float degToDec(float degree, char dir);
 char calcChecksum(std::string str);
 bool isMsgValid(std::string msg);
-int dateTimeToEpoch(unsigned short date[], unsigned short time[]);
+int dateTimeToEpoch(unsigned short date[], float time[]);
 
 #endif	/* CONVERSIONS_H */
